Revert "Feature/responsivite"
